move all url models into the same file
